pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chartgpt"
version = "0.0.2"
authors = [
  { name="Elie Brosset", email="eliebrosset@gmail.com" },
]
description = "ChartGPT is a library for generating charts from text"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pandas>=2.0.0",
    "plotly>=4.0.0",
    "openai>=0.2.0",
    "python-dotenv>=1.0.0",
]
dev = [
    "ruff==0.0.138",
    "pre-commit>=2.17.0,<3.0.0",
]

[project.optional-dependencies]
doc = [
    "mkdocs >=1.1.2,<2.0.0",
    "mkdocs-material >=8.1.4",
    "mkdocs-material-extensions >=1.0.0",
    "mkdocstrings >=0.15.0",
    "mkdocstrings-python >=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/chatgpt/chart"
